requirements.txt
playwright
beautifulsoup4
sqlalchemy
psycopg2-binary
python-dotenv
alembic